requirements-deploy.txt
django==4.0.2
django-environ==0.8.1
psycopg2==2.9.3
pillow==9.0.1
gunicorn==20.1.0
django-storages==1.12.3
boto3==1.21.20
stripe==2.68.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg2==1.19.4
django-cors-headers==3.13.0
whitenoise==6.2.0
